+++ include/Graph.tpp
@@ -4,24 +4,24 @@
 #include "aux.hpp"
 
 template <typename T>
-Graph<T> :: Graph (const uint16_t num_nodes, const T value) :num_nodes_(num_nodes) {
+Graph<T> :: Graph (const uint32_t num_nodes, const T value) :num_nodes_(num_nodes) {
 	
 	paths_weights_.resize(num_nodes);
 
 	// NOTE: In this implementation of Graph the direction does not matter in costs
-	for (uint16_t i = 0; i < num_nodes; i++) {
+	for (uint32_t i = 0; i < num_nodes; i++) {
 		paths_weights_[i] = std::vector<T>(num_nodes - i, value);
 	}
 }
 
 template <typename T>
-uint16_t Graph<T> :: num_nodes() const noexcept {
+uint32_t Graph<T> :: num_nodes() const noexcept {
 	return num_nodes_;
 }
 
 
 template <typename T>
-bool Graph<T> :: is_connected ( const uint16_t first_node, const uint16_t second_node) const noexcept {
+bool Graph<T> :: is_connected ( const uint32_t first_node, const uint32_t second_node) const noexcept {
 
 	T cost_between_nodes = cost(first_node, second_node);
 
@@ -39,12 +39,12 @@ bool Graph<T> :: is_connected ( const uint16_t first_node, const uint16_t second
 }
 
 template <typename T>
-void Graph<T> :: disconnect (const uint16_t first_node, const uint16_t second_node) noexcept {
+void Graph<T> :: disconnect (const uint32_t first_node, const uint32_t second_node) noexcept {
 	connect(first_node, second_node, 0);
 }
 
 template <typename T>
-void Graph<T> :: connect (uint16_t first_node,  uint16_t second_node, const T cost) noexcept {
+void Graph<T> :: connect (uint32_t first_node,  uint32_t second_node, const T cost) noexcept {
 
 	if ( first_node < second_node ) {
 		std::swap(first_node, second_node);
@@ -55,7 +55,7 @@ void Graph<T> :: connect (uint16_t first_node,  uint16_t second_node, const T co
 }
 
 template <typename T>
-T Graph<T> :: cost(uint16_t first_node, uint16_t second_node) const noexcept {
+T Graph<T> :: cost(uint32_t first_node, uint32_t second_node) const noexcept {
 	if ( first_node < second_node ) {
 		std::swap(first_node, second_node);
 	}
